change auth

+++ look/gql/base.py
@@ -6,9 +6,10 @@ from graphene_sqlalchemy import SQLAlchemyObjectType
 
 from graphene_sqlalchemy.converter import convert_sqlalchemy_type
 
+from look.config import Config
 from look.model import Base
 
-from .util import create_gql_model, create_query_field, create_input_class, create_mutation_field, get_instance_by_pk
+from .util import create_gql_model, create_query_field, create_input_class, create_mutation_field, get_instance_by_pk, check_row_by_user_id
 
 def create_base_schema():
     models = {}
@@ -21,6 +22,9 @@ def create_base_schema():
     def resolve_model(self, info, model, **kwargs):
         query = model.get_query(info)
         search = info.context.get('search', None)
+        
+        if 'password' in kwargs:
+            raise Exception("you can't find user with password")
 
         for arg, value in kwargs.items():
             if search:
@@ -47,30 +51,40 @@ def create_base_schema():
             return cls(**{model.__tablename__:instance})
 
     def update_mutate(cls, info, model=None, **kwargs):
-        query = model.get_query(info)
-        model = model._meta.model
-        data = kwargs.get('data', None)
-        if data:
-            data["modified"] = datetime.datetime.utcnow()
-            if 'password' in data: data['password'] = hmac.new(Config.SECRET_KEY.encode(), data['password'].encode(), sha256).hexdigest()
-            
-            instance = get_instance_by_pk(query, model, data)
+        if info.context['auth']['data']:
+            query = model.get_query(info)
+            model = model._meta.model
+            data = kwargs.get('data', None)
+            if data:
+                data["modified"] = datetime.datetime.utcnow()
+                if 'password' in data: data['password'] = hmac.new(Config.SECRET_KEY.encode(), data['password'].encode(), sha256).hexdigest()
+                
+                instance = get_instance_by_pk(query, model, data)
 
-            instance.update(data)
-
-            return cls(**{model.__tablename__:instance.one()})
+                if info.context['auth']['data']['admin'] and check_row_by_user_id(info.context['auth']['data']['user_id'], model, instance):
+                    instance.update(data)
+                    return cls(**{model.__tablename__:instance.one()})
+                else:
+                    raise Exception('PERMISSION DENIED')
+        else:
+            raise Exception(info.context['auth']['description'])
 
     def delete_mutate(cls, info, model=None, **kwargs):
-        query = model.get_query(info)
-        model = model._meta.model
-        data = kwargs.get('data', None)
-        if data:
-            instance = get_instance_by_pk(query, model, data)
-            tmp_instance = instance.one()
+        if info.context['auth']['data']:
+            query = model.get_query(info)
+            model = model._meta.model
+            data = kwargs.get('data', None)
+            if data:
+                instance = get_instance_by_pk(query, model, data)
 
-            instance.delete()
-
-            return cls(**{model.__tablename__:tmp_instance})
+                if info.context['auth']['data']['admin'] and check_row_by_user_id(info.context['auth']['data']['user_id'], model, instance):
+                    tmp_instance = instance.one()
+                    instance.delete()
+                    return cls(**{model.__tablename__:tmp_instance})
+                else:
+                    raise Exception('PERMISSION DENIED')
+        else:
+            raise Exception(info.context['auth']['description'])
 
     query_field = {}
     mutation_field = {}
